stream video on mobile CORE-8463 (#13283)
* wip

* wip

* wip

* wip

* works

* wip

* wtf

* wip

* lol

* morl

* cche

* works but test fails

* fix test

* nicer

* comments

* expand

* one more thing

* fix tests

* feedback

* use adaptive readahead

+++ go/chat/attachments/uploader.go
@@ -400,7 +400,7 @@ func (u *Uploader) upload(ctx context.Context, uid gregor1.UID, convID chat1.Con
 		task := UploadTask{
 			S3Params:       s3params,
 			Filename:       filename,
-			FileSize:       int(finfo.Size()),
+			FileSize:       finfo.Size(),
 			Plaintext:      src,
 			S3Signer:       u.s3signer,
 			ConversationID: convID,
@@ -454,7 +454,7 @@ func (u *Uploader) upload(ctx context.Context, uid gregor1.UID, convID chat1.Con
 			task := UploadTask{
 				S3Params:       previewParams,
 				Filename:       filename,
-				FileSize:       len(pre.Preview),
+				FileSize:       int64(len(pre.Preview)),
 				Plaintext:      newBufReadResetter(pre.Preview),
 				S3Signer:       u.s3signer,
 				ConversationID: convID,
